fix: tests (#15)
This pull request includes changes to update download URLs in extractor
classes and modify the workflow configuration. The most important
changes are grouped by theme below:

### Workflow Configuration:

*
[`.github/workflows/workflow.yml`](diffhunk://#diff-126bf89616b7daa3d14ebc882ad18666aaf1c3dae888c4ba306a66ec80758bc1L79):
Removed the `enable-cache` parameter from the `Install uv` step.

### Extractor Classes:

*
[`src/torah_dl/core/extractors/outorah.py`](diffhunk://#diff-8050ef265bbf3ac9c1dd7797c74e19ca953e0f11a383071020f56bd5ce24f84eL23-R23):
Updated the `download_url` in the `OutorahExtractor` class to a new URL
format.
*
[`src/torah_dl/core/extractors/torahapp.py`](diffhunk://#diff-2dcc74b855b044bfc00230d35ad1118f2c6543bb213316b21224ae1cdf7a29adL36-R36):
Updated the `download_url` in the `TorahAppExtractor` class to a new URL
format.